change app router on projects
